starters/dockerized-django32/full/{{cookiecutter.project_slug}}/requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile base.in
#
aioredis==1.3.1
    # via channels-redis
argon2-cffi==21.3.0
    # via -r base.in
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asgiref==3.3.4
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
async-timeout==4.0.2
    # via
    #   aioredis
    #   redis
attrs==22.1.0
    # via
    #   automat
    #   service-identity
    #   twisted
autobahn==22.7.1
    # via daphne
automat==22.10.0
    # via twisted
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
channels==3.0.4
    # via
    #   -r base.in
    #   channels-redis
channels-redis==3.3.1
    # via -r base.in
charset-normalizer==2.0.10
    # via requests
constantly==15.1.0
    # via twisted
cryptography==38.0.4
    # via
    #   autobahn
    #   pyjwt
    #   pyopenssl
    #   service-identity
daphne==3.0.2
    # via channels
defusedxml==0.7.1
    # via python3-openid
dj-database-url @ git+https://github.com/jazzband/dj-database-url@v1.0.0
    # via -r base.in
django==3.2.16
    # via
    #   -r base.in
    #   channels
    #   dj-database-url
    #   django-allauth
    #   django-browser-reload
    #   django-htmx
    #   django-improved-user
    #   django-lifecycle
    #   django-organizations
    #   django-readers
    #   django-tailwind
    #   django-zen-queries
    #   heroicons
django-allauth==0.51.0
    # via -r base.in
django-browser-reload==1.6.0
    # via django-tailwind
django-environ==0.9.0
    # via -r base.in
django-htmx==1.13.0
    # via -r base.in
django-improved-user==2.0a2
    # via -r base.in
django-lifecycle==1.0.0
    # via -r base.in
django-organizations==2.0.1
    # via -r base.in
django-readers==2.0.0
    # via -r base.in
django-tailwind==3.4.0
    # via -r base.in
django-zen-queries==2.1.0
    # via -r base.in
heroicons[django]==2.1.0
    # via -r base.in
hiredis==2.0.0
    # via aioredis
huey==2.4.3
    # via -r base.in
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.3
    # via
    #   hyperlink
    #   requests
    #   twisted
incremental==22.10.0
    # via twisted
marshmallow==3.14.1
    # via
    #   -r base.in
    #   webargs
msgpack==1.0.4
    # via channels-redis
oauthlib==3.1.1
    # via requests-oauthlib
packaging==21.3
    # via
    #   django-lifecycle
    #   redis
    #   webargs
psycopg2-binary==2.9.5
    # via -r base.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.2.8
    # via service-identity
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.3.0
    # via django-allauth
pyopenssl==22.1.0
    # via twisted
pyparsing==3.0.7
    # via packaging
python-slugify==6.1.2
    # via -r base.in
python3-openid==3.2.0
    # via django-allauth
pytz==2022.6
    # via
    #   -r base.in
    #   django
redis==4.3.5
    # via -r base.in
requests==2.27.1
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via django-allauth
service-identity==21.1.0
    # via twisted
six==1.16.0
    # via
    #   automat
    #   django-organizations
    #   service-identity
sqlparse==0.4.1
    # via django
text-unidecode==1.3
    # via python-slugify
twisted[tls]==22.10.0
    # via daphne
txaio==22.2.1
    # via autobahn
typing-extensions==4.4.0
    # via twisted
urllib3==1.26.8
    # via requests
urlman==2.0.1
    # via django-lifecycle
webargs==8.1.0
    # via -r base.in
zope-interface==5.5.2
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
